Improve git-obs startup time by loading less python modules
